// apps/web/tokens/generated/class-maps.ts
export const directionMap: ResponsiveClassMap<Direction> = {
  base: {
    "row": "flex-row",
    "row-reverse": "flex-row-reverse",
    "col": "flex-col",
    "col-reverse": "flex-col-reverse",
  },
  "sm": {
    "row": "sm:flex-row",
    "row-reverse": "sm:flex-row-reverse",
    "col": "sm:flex-col",
    "col-reverse": "sm:flex-col-reverse",
  },
  "md": {
    "row": "md:flex-row",
    "row-reverse": "md:flex-row-reverse",
    "col": "md:flex-col",
    "col-reverse": "md:flex-col-reverse",
  },
  "lg": {
    "row": "lg:flex-row",
    "row-reverse": "lg:flex-row-reverse",
    "col": "lg:flex-col",
    "col-reverse": "lg:flex-col-reverse",
  },
  "xl": {
    "row": "xl:flex-row",
    "row-reverse": "xl:flex-row-reverse",
    "col": "xl:flex-col",
    "col-reverse": "xl:flex-col-reverse",
  },
  "2xl": {
    "row": "2xl:flex-row",
    "row-reverse": "2xl:flex-row-reverse",
    "col": "2xl:flex-col",
    "col-reverse": "2xl:flex-col-reverse",
  },
};

export const stackAlignMap: ResponsiveClassMap<StackAlign> = {
  base: {
    "start": "items-start",
    "center": "items-center",
    "end": "items-end",
    "stretch": "items-stretch",
    "baseline": "items-baseline",
  },
  "sm": {
    "start": "sm:items-start",
    "center": "sm:items-center",
    "end": "sm:items-end",
    "stretch": "sm:items-stretch",
    "baseline": "sm:items-baseline",
  },
  "md": {
    "start": "md:items-start",
    "center": "md:items-center",
    "end": "md:items-end",
    "stretch": "md:items-stretch",
    "baseline": "md:items-baseline",
  },
  "lg": {
    "start": "lg:items-start",
    "center": "lg:items-center",
    "end": "lg:items-end",
    "stretch": "lg:items-stretch",
    "baseline": "lg:items-baseline",
  },
  "xl": {
    "start": "xl:items-start",
    "center": "xl:items-center",
    "end": "xl:items-end",
    "stretch": "xl:items-stretch",
    "baseline": "xl:items-baseline",
  },
  "2xl": {
    "start": "2xl:items-start",
    "center": "2xl:items-center",
    "end": "2xl:items-end",
    "stretch": "2xl:items-stretch",
    "baseline": "2xl:items-baseline",
  },
};

export const justifyMap: ResponsiveClassMap<Justify> = {
  base: {
    "start": "justify-start",
    "center": "justify-center",
    "end": "justify-end",
    "between": "justify-between",
    "around": "justify-around",
    "evenly": "justify-evenly",
  },
  "sm": {
    "start": "sm:justify-start",
    "center": "sm:justify-center",
    "end": "sm:justify-end",
    "between": "sm:justify-between",
    "around": "sm:justify-around",
    "evenly": "sm:justify-evenly",
  },
  "md": {
    "start": "md:justify-start",
    "center": "md:justify-center",
    "end": "md:justify-end",
    "between": "md:justify-between",
    "around": "md:justify-around",
    "evenly": "md:justify-evenly",
  },
  "lg": {
    "start": "lg:justify-start",
    "center": "lg:justify-center",
    "end": "lg:justify-end",
    "between": "lg:justify-between",
    "around": "lg:justify-around",
    "evenly": "lg:justify-evenly",
  },
  "xl": {
    "start": "xl:justify-start",
    "center": "xl:justify-center",
    "end": "xl:justify-end",
    "between": "xl:justify-between",
    "around": "xl:justify-around",
    "evenly": "xl:justify-evenly",
  },
  "2xl": {
    "start": "2xl:justify-start",
    "center": "2xl:justify-center",
    "end": "2xl:justify-end",
    "between": "2xl:justify-between",
    "around": "2xl:justify-around",
    "evenly": "2xl:justify-evenly",
  },
};

export const stackWrapMap: BaseClassMap<StackWrap> = {
  "nowrap": "flex-nowrap",
  "wrap": "flex-wrap",
  "wrap-reverse": "flex-wrap-reverse",
};

export const columnsMap: ResponsiveClassMap<Columns> = {
  base: {
    "1": "grid-cols-1",
    "2": "grid-cols-2",
    "3": "grid-cols-3",
    "4": "grid-cols-4",
    "5": "grid-cols-5",
    "6": "grid-cols-6",
    "7": "grid-cols-7",
    "8": "grid-cols-8",
    "9": "grid-cols-9",
    "10": "grid-cols-10",
    "11": "grid-cols-11",
    "12": "grid-cols-12",
  },
  "sm": {
    "1": "sm:grid-cols-1",
    "2": "sm:grid-cols-2",
    "3": "sm:grid-cols-3",
    "4": "sm:grid-cols-4",
    "5": "sm:grid-cols-5",
    "6": "sm:grid-cols-6",
    "7": "sm:grid-cols-7",
    "8": "sm:grid-cols-8",
    "9": "sm:grid-cols-9",
    "10": "sm:grid-cols-10",
    "11": "sm:grid-cols-11",
    "12": "sm:grid-cols-12",
  },
  "md": {
    "1": "md:grid-cols-1",
    "2": "md:grid-cols-2",
    "3": "md:grid-cols-3",
    "4": "md:grid-cols-4",
    "5": "md:grid-cols-5",
    "6": "md:grid-cols-6",
    "7": "md:grid-cols-7",
    "8": "md:grid-cols-8",
    "9": "md:grid-cols-9",
    "10": "md:grid-cols-10",
    "11": "md:grid-cols-11",
    "12": "md:grid-cols-12",
  },
  "lg": {
    "1": "lg:grid-cols-1",
    "2": "lg:grid-cols-2",
    "3": "lg:grid-cols-3",
    "4": "lg:grid-cols-4",
    "5": "lg:grid-cols-5",
    "6": "lg:grid-cols-6",
    "7": "lg:grid-cols-7",
    "8": "lg:grid-cols-8",
    "9": "lg:grid-cols-9",
    "10": "lg:grid-cols-10",
    "11": "lg:grid-cols-11",
    "12": "lg:grid-cols-12",
  },
  "xl": {
    "1": "xl:grid-cols-1",
    "2": "xl:grid-cols-2",
    "3": "xl:grid-cols-3",
    "4": "xl:grid-cols-4",
    "5": "xl:grid-cols-5",
    "6": "xl:grid-cols-6",
    "7": "xl:grid-cols-7",
    "8": "xl:grid-cols-8",
    "9": "xl:grid-cols-9",
    "10": "xl:grid-cols-10",
    "11": "xl:grid-cols-11",
    "12": "xl:grid-cols-12",
  },
  "2xl": {
    "1": "2xl:grid-cols-1",
    "2": "2xl:grid-cols-2",
    "3": "2xl:grid-cols-3",
    "4": "2xl:grid-cols-4",
    "5": "2xl:grid-cols-5",
    "6": "2xl:grid-cols-6",
    "7": "2xl:grid-cols-7",
    "8": "2xl:grid-cols-8",
    "9": "2xl:grid-cols-9",
    "10": "2xl:grid-cols-10",
    "11": "2xl:grid-cols-11",
    "12": "2xl:grid-cols-12",
  },
};
// ...
export const minItemWidthMap: BaseClassMap<MinItemWidth> = {
  "xs": "grid-auto-xs",
  "sm": "grid-auto-sm",
  "md": "grid-auto-md",
  "lg": "grid-auto-lg",
};